requirements.txt
attrs==23.1.0
black==22.3.0
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
contourpy==1.1.1
coverage==7.3.1
cycler==0.12.1
fonttools==4.43.1
idna==3.4
iniconfig==2.0.0
kiwisolver==1.4.5
matplotlib==3.8.0
mypy-extensions==1.0.0
numpy==1.26.0
packaging==23.1
pandas==2.1.1
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.10.0
pluggy==1.3.0
py==1.11.0
pyodbc==4.0.39
pyparsing==3.1.1
pypyodbc==1.3.6
pytest==7.1.3
pytest-cov==4.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.26.0
ruff==0.0.284
six==1.16.0
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.16
